// apps/backend/src/app/modules/permission/enum/apps.permission.enum.ts
enum AppsPermission {
  VIEW_ALL_APP = '8e3ee9d8-f109-4ae3-9cf2-d33ed05eee5a',
  EDIT_ALL_APP = '3c714bf7-8f32-4fff-a820-e69f0fa291f9',
  CREATE_ALL_APP = '23a3f8e2-ffb6-4102-aa17-1ea6f63aea8b',
  DELETE_ALL_APP_VERSION = '6132981e-7795-4540-84be-4d7c3ec531cc',
  CREATE_ALL_APP_VERSION = '1cda4ec8-a441-43b3-b526-da21c52237ef',

  VIEW_APP = '98e3ccbf-9450-4809-8bfd-f527b1bfe8fc',
  CREATE_APP = 'ecaeafeb-0b11-4bcb-af5c-9a4f15aee1de',
  EDIT_APP = 'c710dbe5-d9db-421d-b4b5-7abd04a6cf9d',
  DELETE_APP_VERSION = 'b193735b-1918-47a4-9878-1209430d434f',
  CREATE_APP_VERSION = '3f783226-f67e-4062-8566-cbf77c7f8f88',
}

export default AppsPermission;